chore: move event-sourced entity codegen into its own file for consistency (#728)
