+++ codegen/java-gen/src/main/scala/com/lightbend/akkasls/codegen/java/EntityServiceSourceGenerator.scala
@@ -75,8 +75,7 @@ object EntityServiceSourceGenerator {
       implSourcePath.getParent.toFile.mkdirs()
       Files.write(
         implSourcePath,
-        source(service, entity, packageName, implClassName, interfaceClassName, entity.entityType).getBytes(
-          Charsets.UTF_8))
+        source(service, entity, packageName, implClassName, interfaceClassName).getBytes(Charsets.UTF_8))
     }
 
     val routerClassName = entity.routerName
@@ -131,254 +130,20 @@ object EntityServiceSourceGenerator {
       entity: ModelBuilder.Entity,
       packageName: String,
       className: String,
-      interfaceClassName: String,
-      entityType: String): String = {
+      interfaceClassName: String): String = {
     entity match {
       case eventSourcedEntity: EventSourcedEntity =>
-        eventSourcedEntitySource(service, eventSourcedEntity, packageName, className, interfaceClassName)
+        EventSourcedEntitySourceGenerator.eventSourcedEntitySource(
+          service,
+          eventSourcedEntity,
+          packageName,
+          className,
+          interfaceClassName)
       case valueEntity: ValueEntity =>
         ValueEntitySourceGenerator.valueEntitySource(service, valueEntity, packageName, className)
       case replicatedEntity: ReplicatedEntity =>
         ReplicatedEntitySourceGenerator.replicatedEntitySource(service, replicatedEntity, packageName, className)
     }
-  }
-
-  private[codegen] def eventSourcedEntityRouter(
-      service: ModelBuilder.EntityService,
-      entity: ModelBuilder.EventSourcedEntity,
-      packageName: String,
-      className: String): String = {
-
-    val imports = generateCommandImports(
-      service.commands,
-      entity.state,
-      packageName,
-      otherImports = Seq(
-        "com.akkaserverless.javasdk.eventsourcedentity.CommandContext",
-        "com.akkaserverless.javasdk.eventsourcedentity.EventSourcedEntity",
-        "com.akkaserverless.javasdk.impl.eventsourcedentity.EventSourcedEntityRouter"))
-
-    val stateType = entity.state.fqn.fullName
-
-    val eventCases = {
-      if (entity.events.isEmpty)
-        List(s"throw new EventSourcedEntityRouter.EventHandlerNotFound(event.getClass());")
-      else
-        entity.events.zipWithIndex.map { case (evt, i) =>
-          val eventType = evt.fqn.fullName
-          s"""|${if (i == 0) "" else "} else "}if (event instanceof $eventType) {
-              |  return entity().${lowerFirst(evt.fqn.name)}(state, ($eventType) event);""".stripMargin
-        }.toSeq :+
-        s"""|} else {
-          |  throw new EventSourcedEntityRouter.EventHandlerNotFound(event.getClass());
-          |}""".stripMargin
-    }
-
-    val commandCases = service.commands
-      .map { cmd =>
-        val methodName = cmd.name
-        val inputType = cmd.inputType.fullName
-        s"""|case "$methodName":
-            |  return entity().${lowerFirst(methodName)}(state, ($inputType) command);
-            |""".stripMargin
-      }
-
-    s"""package $packageName;
-        |
-        |${writeImports(imports)}
-        |
-        |$managedComment
-        |
-        |/**
-        | * An event sourced entity handler that is the glue between the Protobuf service <code>${service.fqn.name}</code>
-        | * and the command and event handler methods in the <code>${entity.fqn.name}</code> class.
-        | */
-        |public class ${className}Router extends EventSourcedEntityRouter<$stateType, ${entity.fqn.name}> {
-        |
-        |  public ${className}Router(${entity.fqn.name} entity) {
-        |    super(entity);
-        |  }
-        |
-        |  @Override
-        |  public $stateType handleEvent($stateType state, Object event) {
-        |    ${Format.indent(eventCases, 4)}
-        |  }
-        |
-        |  @Override
-        |  public EventSourcedEntity.Effect<?> handleCommand(
-        |      String commandName, $stateType state, Object command, CommandContext context) {
-        |    switch (commandName) {
-        |
-        |      ${Format.indent(commandCases, 6)}
-        |
-        |      default:
-        |        throw new EventSourcedEntityRouter.CommandHandlerNotFound(commandName);
-        |    }
-        |  }
-        |}
-        |""".stripMargin
-
-  }
-
-  private[codegen] def eventSourcedEntityProvider(
-      service: ModelBuilder.EntityService,
-      entity: ModelBuilder.EventSourcedEntity,
-      packageName: String,
-      className: String): String = {
-    val relevantTypes = {
-      service.commands.flatMap { cmd =>
-        cmd.inputType :: cmd.outputType :: Nil
-      }.toSeq :+ entity.state.fqn
-    }
-
-    implicit val imports: Imports = generateImports(
-      relevantTypes ++ relevantTypes.map(_.descriptorImport),
-      packageName,
-      otherImports = Seq(
-        "com.akkaserverless.javasdk.eventsourcedentity.EventSourcedEntityContext",
-        "com.akkaserverless.javasdk.eventsourcedentity.EventSourcedEntityOptions",
-        "com.akkaserverless.javasdk.eventsourcedentity.EventSourcedEntityProvider",
-        "com.google.protobuf.Descriptors",
-        "java.util.function.Function"))
-
-    val descriptors =
-      (collectRelevantTypes(relevantTypes, service.fqn)
-        .map(d =>
-          s"${d.parent.javaOuterClassname}.getDescriptor()") :+ s"${service.fqn.parent.javaOuterClassname}.getDescriptor()").distinct.sorted
-
-    s"""package $packageName;
-        |
-        |${writeImports(imports)}
-        |
-        |$managedComment
-        |
-        |/**
-        | * An event sourced entity provider that defines how to register and create the entity for
-        | * the Protobuf service <code>${service.fqn.name}</code>.
-        | *
-        | * Should be used with the <code>register</code> method in {@link com.akkaserverless.javasdk.AkkaServerless}.
-        | */
-        |public class ${className}Provider implements EventSourcedEntityProvider<${entity.state.fqn.fullName}, $className> {
-        |
-        |  private final Function<EventSourcedEntityContext, $className> entityFactory;
-        |  private final EventSourcedEntityOptions options;
-        |
-        |  /** Factory method of ${className}Provider */
-        |  public static ${className}Provider of(Function<EventSourcedEntityContext, $className> entityFactory) {
-        |    return new ${className}Provider(entityFactory, EventSourcedEntityOptions.defaults());
-        |  }
-        |
-        |  private ${className}Provider(
-        |      Function<EventSourcedEntityContext, $className> entityFactory,
-        |      EventSourcedEntityOptions options) {
-        |    this.entityFactory = entityFactory;
-        |    this.options = options;
-        |  }
-        |
-        |  @Override
-        |  public final EventSourcedEntityOptions options() {
-        |    return options;
-        |  }
-        |
-        |  public final ${className}Provider withOptions(EventSourcedEntityOptions options) {
-        |    return new ${className}Provider(entityFactory, options);
-        |  }
-        |
-        |  @Override
-        |  public final Descriptors.ServiceDescriptor serviceDescriptor() {
-        |    return ${typeName(service.fqn.descriptorImport)}.getDescriptor().findServiceByName("${service.fqn.name}");
-        |  }
-        |
-        |  @Override
-        |  public final String entityType() {
-        |    return "${entity.entityType}";
-        |  }
-        |
-        |  @Override
-        |  public final ${className}Router newRouter(EventSourcedEntityContext context) {
-        |    return new ${className}Router(entityFactory.apply(context));
-        |  }
-        |
-        |  @Override
-        |  public final Descriptors.FileDescriptor[] additionalDescriptors() {
-        |    return new Descriptors.FileDescriptor[] {
-        |      ${Format.indent(descriptors.mkString(",\n"), 6)}
-        |    };
-        |  }
-        |}
-        |""".stripMargin
-
-  }
-
-  private[codegen] def eventSourcedEntitySource(
-      service: ModelBuilder.EntityService,
-      entity: ModelBuilder.EventSourcedEntity,
-      packageName: String,
-      className: String,
-      interfaceClassName: String): String = {
-    val messageTypes = service.commands.toSeq
-      .flatMap(command => Seq(command.inputType, command.outputType)) ++
-      entity.events.map(_.fqn) :+ entity.state.fqn
-
-    val imports = generateCommandImports(
-      service.commands,
-      entity.state,
-      packageName,
-      Seq(
-        "com.akkaserverless.javasdk.eventsourcedentity.EventSourcedEntityContext",
-        "com.akkaserverless.javasdk.eventsourcedentity.EventSourcedEntity",
-        "com.akkaserverless.javasdk.eventsourcedentity.EventSourcedEntity.Effect"))
-
-    val commandHandlers =
-      service.commands
-        .map { command =>
-          s"""|@Override
-              |public Effect<${qualifiedType(command.outputType)}> ${lowerFirst(command.name)}(${qualifiedType(
-            entity.state.fqn)} currentState, ${qualifiedType(command.inputType)} ${lowerFirst(command.inputType.name)}) {
-              |  return effects().error("The command handler for `${command.name}` is not implemented, yet");
-              |}
-              |""".stripMargin
-        }
-
-    val eventHandlers =
-      entity match {
-        case ModelBuilder.EventSourcedEntity(_, _, _, events) =>
-          events.map { event =>
-            s"""|@Override
-                |public ${qualifiedType(entity.state.fqn)} ${lowerFirst(event.fqn.name)}(${qualifiedType(
-              entity.state.fqn)} currentState, ${qualifiedType(event.fqn)} ${lowerFirst(event.fqn.name)}) {
-                |  throw new RuntimeException("The event handler for `${event.fqn.name}` is not implemented, yet");
-                |}""".stripMargin
-          }
-      }
-
-    s"""package $packageName;
-       |
-       |${writeImports(imports)}
-       |
-       |$unmanagedComment
-       |
-       |/** An event sourced entity. */
-       |public class $className extends ${interfaceClassName} {
-       |
-       |  @SuppressWarnings("unused")
-       |  private final String entityId;
-       |
-       |  public $className(EventSourcedEntityContext context) {
-       |    this.entityId = context.entityId();
-       |  }
-       |
-       |  @Override
-       |  public ${qualifiedType(entity.state.fqn)} emptyState() {
-       |    throw new UnsupportedOperationException("Not implemented yet, replace with your empty entity state");
-       |  }
-       |
-       |  ${Format.indent(commandHandlers, num = 2)}
-       |
-       |  ${Format.indent(eventHandlers, num = 2)}
-       |
-       |}
-       |""".stripMargin
   }
 
   private[codegen] def interfaceSource(
@@ -389,7 +154,12 @@ object EntityServiceSourceGenerator {
       mainPackageName: String): String =
     entity match {
       case eventSourcedEntity: ModelBuilder.EventSourcedEntity =>
-        abstractEventSourcedEntity(service, eventSourcedEntity, packageName, className, mainPackageName)
+        EventSourcedEntitySourceGenerator.abstractEventSourcedEntity(
+          service,
+          eventSourcedEntity,
+          packageName,
+          className,
+          mainPackageName)
       case valueEntity: ModelBuilder.ValueEntity =>
         ValueEntitySourceGenerator.abstractValueEntity(service, valueEntity, packageName, className, mainPackageName)
       case replicatedEntity: ReplicatedEntity =>
@@ -408,7 +178,7 @@ object EntityServiceSourceGenerator {
       className: String): String = {
     entity match {
       case entity: ModelBuilder.EventSourcedEntity =>
-        EntityServiceSourceGenerator.eventSourcedEntityRouter(service, entity, packageName, className)
+        EventSourcedEntitySourceGenerator.eventSourcedEntityRouter(service, entity, packageName, className)
       case entity: ValueEntity =>
         ValueEntitySourceGenerator.valueEntityRouter(service, entity, packageName, className)
       case entity: ReplicatedEntity =>
@@ -423,61 +193,16 @@ object EntityServiceSourceGenerator {
       className: String): String = {
     entity match {
       case eventSourcedEntity: ModelBuilder.EventSourcedEntity =>
-        eventSourcedEntityProvider(service, eventSourcedEntity, packageName, className)
+        EventSourcedEntitySourceGenerator.eventSourcedEntityProvider(
+          service,
+          eventSourcedEntity,
+          packageName,
+          className)
       case valueEntity: ValueEntity =>
         ValueEntitySourceGenerator.valueEntityProvider(service, valueEntity, packageName, className)
       case replicatedEntity: ReplicatedEntity =>
         ReplicatedEntitySourceGenerator.replicatedEntityProvider(service, replicatedEntity, packageName, className)
     }
-  }
-
-  private[codegen] def abstractEventSourcedEntity(
-      service: ModelBuilder.EntityService,
-      entity: ModelBuilder.EventSourcedEntity,
-      packageName: String,
-      className: String,
-      mainPackageName: String): String = {
-    val imports = generateCommandImports(
-      service.commands,
-      entity.state,
-      packageName,
-      Seq(
-        "com.akkaserverless.javasdk.eventsourcedentity.EventSourcedEntity",
-        s"$mainPackageName.Components",
-        s"$mainPackageName.ComponentsImpl"))
-
-    val commandHandlers = service.commands.map { command =>
-      s"""|/** Command handler for "${command.name}". */
-          |public abstract Effect<${qualifiedType(command.outputType)}> ${lowerFirst(command.name)}(${qualifiedType(
-        entity.state.fqn)} currentState, ${qualifiedType(command.inputType)} ${lowerFirst(command.inputType.name)});
-         |""".stripMargin
-    }
-
-    val eventHandlers = entity.events.map { event =>
-      s"""|/** Event handler for "${event.fqn.name}". */
-          |public abstract ${qualifiedType(entity.state.fqn)} ${lowerFirst(event.fqn.name)}(${qualifiedType(
-        entity.state.fqn)} currentState, ${qualifiedType(event.fqn)} ${lowerFirst(event.fqn.name)});
-         |""".stripMargin
-    }
-
-    s"""package $packageName;
-        |
-        |${writeImports(imports)}
-        |
-        |$managedComment
-        |
-        |/** An event sourced entity. */
-        |public abstract class Abstract${className} extends EventSourcedEntity<${qualifiedType(entity.state.fqn)}> {
-        |
-        |  protected final Components components() {
-        |    return new ComponentsImpl(commandContext());
-        |  }
-        |
-        |  ${Format.indent(commandHandlers, num = 2)}
-        |
-        |  ${Format.indent(eventHandlers, num = 2)}
-        |
-        |}""".stripMargin
   }
 
   private[codegen] def integrationTestSource(
